requirements.txt
flake8==3.8.3
importlib-metadata==1.7.0
mccabe==0.6.1
mysql-connector-python==8.0.21
protobuf==3.12.2
pycodestyle==2.6.0
pyflakes==2.2.0
six==1.15.0
zipp==3.1.0
